Refactor filtertracker reducer - Rename tracklist reducer - Move to hooks GetTracksHook function
