+++ src/hooks/index.js
@@ -1,17 +1,42 @@
+import { useEffect, useReducer } from 'react';
+import ACTIONS from '../actions';
 import API from '../Api';
-
-const { useEffect, useState } = require('react');
-
-const GetTracksHook = artistName => {
-  const [track, setTrack] = useState([])
-  // useEffect(() => {
-  API.getArtist(artistName).then(data => {
-    setTrack(data.tracks);
-  });
-  // }, []);
+import trackReducer from '../reducers/trackReducer';
 
 
-  return track;
+const addToObject = (objectArray = []) => {
+  const responseArray = []
+
+  objectArray.forEach(elem => {
+    const response = {
+      id: elem.id,
+      name: elem.name,
+      album: elem.album.name,
+      year: new Date(elem.album.release_date).getFullYear(),
+      cover: elem.album.images[1].url
+    }
+    responseArray.push(response)
+  })
+  return responseArray;
+}
+
+const GetTracksHook = () => {
+  const store = {
+    tracks: [],
+    filter: 'All',
+  };
+  const [state, dispatch] = useReducer(trackReducer, store);
+
+  useEffect(() => {
+    API.getArtist('Van Halen').then(data => {
+      dispatch({
+        type: ACTIONS.INITIALIZE,
+        payload: addToObject(data.tracks.items)
+      })
+    });
+  }, [])
+
+  return state;
 };
 
 export default GetTracksHook;
